speed up AES finder.
